Extract state machine to separate file
- Create state_machine.h/cpp with checkMachineState() function
- Move STEAM_TEMP_DETECTION constant to state_machine.h
- Move cold start tracking variables to state_machine.cpp as static
- Reduces main.cpp from 353 to 220 lines

This improves separation of concerns and makes the state machine
logic easier to understand and modify independently.
