+++ include/state_machine.h
@@ -1,0 +1,6 @@
+#pragma once
+
+// Steam detection temperature threshold
+#define STEAM_TEMP_DETECTION 120
+
+void checkMachineState();
